requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.2
psycopg2-binary==2.8.5